requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
flask-smorest==0.42.3
python-dotenv==1.0.0
PyJWT==2.8.0
camelot-py==0.11.0
marshmallow==3.19.0
Werkzeug==2.3.0
requests==2.31.0
Flask-CORS==4.0.0
